refactor(front): add component folder
