+++ frontend/src/App.css
@@ -36,3 +36,15 @@
     transform: rotate(360deg);
   }
 }
+
+.site-layout .site-layout-background {
+  background: #fff;
+}
+/* [data-theme="dark"] .site-layout .site-layout-background {
+  background: #141414;
+} */
+
+.site-card-wrapper {
+  padding: 30px;
+  background: #ececec;
+}
