--- frontend/src/components/Detail.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { Layout, Menu, Breadcrumb } from "antd";
-import { Content, Sider } from "antd/lib/layout/layout";
-
-const { SubMenu } = Detail;
-
-const Detail = () => {
-  return <>Detail</>;
-};
-
-export default Detail;
