+++ frontend/src/components/auth/Join.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Join = () => {
   return <div>Join</div>;
 };
