+++ frontend/src/components/auth/Login.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Login = () => {
   return <div>Login</div>;
 };
