+++ frontend/src/components/meet/Detail.js
@@ -1,0 +1,7 @@
+const { SubMenu } = Detail;
+
+const Detail = () => {
+  return <>Detail</>;
+};
+
+export default Detail;
