+++ frontend/src/components/spinner/Spinner.js
@@ -1,0 +1,10 @@
+import "antd/dist/antd.css";
+import {Spin} from "antd";
+
+export const GrowingSpinner = (
+  <>
+    <Row className="d-flex justify-content-center m-5">
+      <Spin />
+    </Row>
+  </>
+);
